Derive study and visit identifiers from PatientID
This relies on an underscore convention being used when PatientIDs are
generated on the scanner console.

The (orthanc) study ID -> PatientID pairs are used to populate the
selection list widget and stored as a property to make lookup
easy (.selected returns only values, ie. orthanc IDs).

It would be possible to rely on more than PatientID by performing an
additional request to the /studies endpoint (or by storing more
information upfront).