Descend into subdirectory of orthanc export
We export Orthanc zipfile and unpack it into a directory named with
the orthanc study id. The zipfile (and, consequently, unpacked folder)
is named by Orthanc based on DICOM tags, "<patientID> <patientName>".

This is the directory we want to run ICF utils on. By ICF utils logic,
the top level directory will be replaced with the given study/visit
identifiers (which we currently derive from the patientID field only)
when creating a tarball.

This effectively means that we will not be writing patientName into
the generated paths.